test: dedupe _FakeReadyPage/_FakePage into conftest.FakeEvaluatePage
test_browser.py and test_resy_url_match.py each defined a byte-for-byte
identical fake page class (queue of scripted page.evaluate() results,
raising when the scripted value is an exception) under different names,
to drive wait_for_page_ready's retry loop and ResyUrlMatch.update's two
sequential evaluate calls respectively. Share one FakeEvaluatePage in
tests/conftest.py instead, mirroring the existing run_async precedent.

No behavior change: test_browser.py imports it under its old local name
via `as _FakeReadyPage` so its call sites are untouched, and
test_resy_url_match.py's single call site is updated to the shared name.
